+++ standalone/include/vpn/standalone/config.h
@@ -49,6 +49,7 @@ struct VpnStandaloneConfig {
     ag::LogLevel loglevel = ag::LOG_LEVEL_INFO;
     ag::VpnMode mode = ag::VPN_MODE_GENERAL;
     bool killswitch_enabled = false;
+    bool post_quantum_group_enabled = false;
     std::string log_file_path;
     std::string exclusions;
     std::optional<std::string> ssl_session_storage_path;